refactor: rename `google.mts` to `rank.mts`
